backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.29,<1
httpx>=0.27,<1
python-dotenv>=1.0,<2
starlette>=0.37,<1
itsdangerous>=2.2,<3
pydantic>=2.7,<3
python-multipart>=0.0.9,<1
google-cloud-storage>=2.16,<3
